Map-making

Map-making

Harlequin includes a routine that applies the destriping algorithm to time-ordered data, producing a maximum-likelihood map of the intensity and polarization signal.

The destriping algorithm used in the code is described in the paper Destiping CMB temperature and polarization maps, Kurki-Suonio et al., A/A 506, 1511-1539 (2009), and the source code closely follows the terminology introduced in that paper.

The destriping algorithm is effective in removing 1/f noise originated within the detectors of an instrument, provided that the noise is uncorrelated among detectors. It requires the user to specify a baseline, i.e., the maximum time span in which the noise can be assumed to be uncorrelated (i.e., white).

Since the destriper is effective only when much data is available, it is often the case that the input data to be fed to the algorithm is larger than the available memory on a machine. In this case, the destriper can take advantage of a distributed-memory environment and of the MPI libraries.

High-level functions

Harlequin.destripe! — Function.

destripe!(obs_list, baselines, destriping_data; threshold, max_iterations, comm, unseen, callback, use_preconditioner) where {T <: Number, O <: Healpix.Order}

Apply the destriping algorithm to the TOD in obs_list. The result is returned in baselines (set of baselines) and destriping_data (I/Q/U maps, weight map, M_p matrices).

The parameters must satisfy the following constraints:

obs_list must be an array of N Observation objects;

baselines must be an array of N RunLengthArray objects; each of them must have all the values set to zero; the length of each run specifies the length of the baseline in terms of the samples in the TOD in obs_list

destriping_data must be a DestripingData object. It does not need to have been initialized using reset_maps!.

The optional keyword have the following meaning and default value:

max_iterations is the maximum number of allowed iterations for the Conjugated Gradient algorithm. The default is 1000.

comm is a MPI communicator object, or nothing is MPI is not used.

unseen is the value to be used for pixels in the map that have not been observed. The default is missing; other sensible values are nothing and NaN.

callback is either nothing (the default) or a function that is called before the CG algorithm starts, and then once every iteration. It can be used to monitor the progress of the destriper, or to save intermediate results in a database or a file. The function must accept the following parameters:

Iteration number, starting from 1

Maximum number of iterations, equal to max_iterations

Current stopping factor

The value of destriping_data

The function uses Julia's Logging module. Therefore, you can enable the output of diagnostic messages, like in the following example:

global_logger(ConsoleLogger(stderr, Debug))

# This command is going to produce a lot of output…
destripe!(...)

source

Low-level functions